@@ -1,4 +1,3 @@
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_circular_chart/src/circular_chart.dart';
 import 'package:flutter_circular_chart/src/entry.dart';
@@ -25,9 +24,9 @@ enum SegmentEdgeStyle {
 
 class AnimatedCircularChart extends StatefulWidget {
   AnimatedCircularChart({
-    Key key,
-    @required this.size,
-    @required this.initialChartData,
+    Key? key,
+    required this.size,
+    required this.initialChartData,
     this.chartType = CircularChartType.Radial,
     this.duration = _kDuration,
     this.percentageValues = false,
@@ -76,7 +75,7 @@ class AnimatedCircularChart extends StatefulWidget {
   /// be automatically calculated to accommodate all the data.
   ///
   /// Has no effect in [CircularChartType.Pie] charts.
-  final double holeRadius;
+  final double? holeRadius;
 
   /// The chart gets drawn and animates clockwise from [startAngle], defaulting to the
   /// top/center point or -90.0. In terms of a clock face these would be:
@@ -92,13 +91,13 @@ class AnimatedCircularChart extends StatefulWidget {
   /// in the center of the chart's hole.
   ///
   /// See also [labelStyle] which is used to render the label.
-  final String holeLabel;
+  final String? holeLabel;
 
   /// The style used when rendering the [holeLabel].
   ///
   /// Defaults to the active [ThemeData]'s
   /// [ThemeData.textTheme.body2] text style.
-  final TextStyle labelStyle;
+  final TextStyle? labelStyle;
 
   /// The type of segment edges to be drawn.
   ///
@@ -118,10 +117,10 @@ class AnimatedCircularChart extends StatefulWidget {
     assert(context != null);
     assert(nullOk != null);
 
-    final AnimatedCircularChartState result = context
-        .ancestorStateOfType(const TypeMatcher<AnimatedCircularChartState>());
+    final AnimatedCircularChartState? result = context
+        .findAncestorStateOfType<AnimatedCircularChartState>();
 
-    if (nullOk || result != null) return result;
+    if (nullOk && result != null) return result;
 
     throw new FlutterError(
         'AnimatedCircularChart.of() called with a context that does not contain a AnimatedCircularChart.\n'
@@ -152,8 +151,8 @@ class AnimatedCircularChart extends StatefulWidget {
 /// ```
 class AnimatedCircularChartState extends State<AnimatedCircularChart>
     with TickerProviderStateMixin {
-  CircularChartTween _tween;
-  AnimationController _animation;
+  late CircularChartTween _tween;
+  late AnimationController _animation;
   final Map<String, int> _stackRanks = <String, int>{};
   final Map<String, int> _entryRanks = <String, int>{};
   final TextPainter _labelPainter = new TextPainter();
@@ -208,17 +207,17 @@ class AnimatedCircularChartState extends State<AnimatedCircularChart>
 
   void _assignRanks(List<CircularStackEntry> data) {
     for (CircularStackEntry stackEntry in data) {
-      _stackRanks.putIfAbsent(stackEntry.rankKey, () => _stackRanks.length);
+      _stackRanks.putIfAbsent(stackEntry.rankKey ?? "", () => _stackRanks.length);
       for (CircularSegmentEntry entry in stackEntry.entries) {
-        _entryRanks.putIfAbsent(entry.rankKey, () => _entryRanks.length);
+        _entryRanks.putIfAbsent(entry.rankKey ?? "", () => _entryRanks.length);
       }
     }
   }
 
   void _updateLabelPainter() {
     if (widget.holeLabel != null) {
-      TextStyle _labelStyle =
-          widget.labelStyle ?? Theme.of(context).textTheme.body2;
+      TextStyle? _labelStyle =
+          widget.labelStyle ?? Theme.of(context).textTheme.bodyMedium ;
       _labelPainter
         ..text = new TextSpan(style: _labelStyle, text: widget.holeLabel)
         ..textDirection = Directionality.of(context)
@@ -259,7 +258,7 @@ class AnimatedCircularChartState extends State<AnimatedCircularChart>
       size: widget.size,
       painter: new AnimatedCircularChartPainter(
         _tween.animate(_animation),
-        widget.holeLabel != null ? _labelPainter : null,
+         widget.holeLabel != null ? _labelPainter : null,
       ),
     );
   }
