@@ -19,20 +19,20 @@ class CircularChart {
   final CircularChartType chartType;
   final SegmentEdgeStyle edgeStyle;
 
-  factory CircularChart.empty({@required CircularChartType chartType}) {
+  factory CircularChart.empty({required CircularChartType chartType}) {
     return new CircularChart(<CircularChartStack>[], chartType);
   }
 
   factory CircularChart.fromData({
-    @required Size size,
-    @required List<CircularStackEntry> data,
-    @required CircularChartType chartType,
-    @required bool percentageValues,
-    @required double startAngle,
-    Map<String, int> stackRanks,
-    Map<String, int> entryRanks,
-    double holeRadius,
-    SegmentEdgeStyle edgeStyle,
+    required Size size,
+    required List<CircularStackEntry> data,
+    required CircularChartType chartType,
+    required bool percentageValues,
+    required double startAngle,
+    Map<String, int>? stackRanks,
+    Map<String, int>? entryRanks,
+    double? holeRadius,
+    SegmentEdgeStyle? edgeStyle,
   }) {
     final double _holeRadius = holeRadius ?? size.width / (2 + data.length);
     final double stackDistance =
@@ -43,9 +43,9 @@ class CircularChart {
     List<CircularChartStack> stacks = new List<CircularChartStack>.generate(
       data.length,
       (i) => new CircularChartStack.fromData(
-            stackRanks[data[i].rankKey] ?? i,
+            stackRanks?[data[i].rankKey] ?? i,
             data[i].entries,
-            entryRanks,
+            entryRanks ?? Map<String, int>(),
             percentageValues,
             startRadius + i * stackDistance,
             stackWidth,
@@ -53,7 +53,7 @@ class CircularChart {
           ),
     );
 
-    return new CircularChart(stacks, chartType, edgeStyle: edgeStyle);
+    return new CircularChart(stacks, chartType, edgeStyle: edgeStyle ?? SegmentEdgeStyle.flat);
   }
 }
 
@@ -68,7 +68,7 @@ class CircularChartTween extends Tween<CircularChart> {
   @override
   CircularChart lerp(double t) => new CircularChart(
         _stacksTween.lerp(t),
-        begin.chartType,
-        edgeStyle: end.edgeStyle,
+        begin?.chartType ?? CircularChartType.Radial,
+        edgeStyle: end?.edgeStyle ?? SegmentEdgeStyle.flat ,
       );
 }
