@@ -10,7 +10,7 @@ class AnimatedCircularChartPainter extends CustomPainter {
       : super(repaint: animation);
 
   final Animation<CircularChart> animation;
-  final TextPainter labelPainter;
+  final TextPainter? labelPainter;
 
   @override
   void paint(Canvas canvas, Size size) {
@@ -40,7 +40,7 @@ class CircularChartPainter extends CustomPainter {
 
 const double _kRadiansPerDegree = Math.pi / 180;
 
-void _paintLabel(Canvas canvas, Size size, TextPainter labelPainter) {
+void _paintLabel(Canvas canvas, Size size, TextPainter? labelPainter) {
   if (labelPainter != null) {
     labelPainter.paint(
       canvas,
@@ -63,16 +63,16 @@ void _paintChart(Canvas canvas, Size size, CircularChart chart) {
 
   for (final CircularChartStack stack in chart.stacks) {
     for (final segment in stack.segments) {
-      segmentPaint.color = segment.color;
-      segmentPaint.strokeWidth = stack.width;
+      segmentPaint.color = segment.color ?? Colors.black;
+      segmentPaint.strokeWidth = stack.width ?? 1;
 
       canvas.drawArc(
         new Rect.fromCircle(
           center: new Offset(size.width / 2, size.height / 2),
-          radius: stack.radius,
+          radius: stack.radius ?? 0,
         ),
-        stack.startAngle * _kRadiansPerDegree,
-        segment.sweepAngle * _kRadiansPerDegree,
+        (stack.startAngle ?? 0) * _kRadiansPerDegree,
+        (segment.sweepAngle ?? 0) * _kRadiansPerDegree,
         chart.chartType == CircularChartType.Pie,
         segmentPaint,
       );
