@@ -7,8 +7,8 @@ class CircularChartSegment extends MergeTweenable<CircularChartSegment> {
   CircularChartSegment(this.rank, this.sweepAngle, this.color);
 
   final int rank;
-  final double sweepAngle;
-  final Color color;
+  final double? sweepAngle;
+  final Color? color;
 
   @override
   CircularChartSegment get empty => new CircularChartSegment(rank, 0.0, color);
@@ -40,6 +40,11 @@ class CircularChartSegmentTween extends Tween<CircularChartSegment> {
   }
 
   @override
-  CircularChartSegment lerp(double t) =>
-      CircularChartSegment.lerp(begin, end, t);
+  CircularChartSegment lerp(double t) {
+    if (begin == null || end == null) {
+      return CircularChartSegment(begin?.rank ?? 0, null, null);
+    } 
+    return CircularChartSegment.lerp(begin!, end!, t); 
+  }
+      
 }
